Login refactored
